selectExploreSectionFromTask and selectRecommendedTasks
