+++ src/components/my_plan/my_plan_connected_component.tsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { TaskListItemActions } from '../tasks/task_list_item';
 import { Store } from '../../stores';
 import { MyPlanComponent, MyPlanProps } from './my_plan';
-import { selectRecommendedTasks } from '../../selectors/tasks';
+import { selectRecommendedTasks } from '../../selectors/tasks/select_recommended_tasks';
 import { Id, AddToSavedListAction, addToSavedList } from '../../stores/tasks';
 import { selectSavedTasks } from '../../selectors/tasks/select_saved_tasks';
 import { selectCompletedTasks } from '../../selectors/tasks/select_completed_tasks';
